refactor(db): Chunk location repository implementation
Refactors the monolithic `LocationRepository` implementation into the chunked pattern, where each method and its corresponding tests reside in a separate file. This significantly improves code modularity, maintainability, and navigability.

The old `location_repository_impl.rs` and `location_repository_batch_impl.rs` files have been replaced by a new `location_repository` directory containing the chunked implementation. Tests have been co-located with the code they verify.

Components affected:
- `banking-db-postgres`: Refactored `LocationRepository` and its tests.
- `banking-db-postgres`: Updated module structure in `person/mod.rs`.

Benefits:
- Improved Code Organization: Each method is now in its own file, making the codebase easier to understand.
- Enhanced Maintainability: Changes to a specific method are isolated to a single file.
- Co-located Tests: Tests are now alongside the implementation, improving context and ease of testing.
